requirements.txt
# Core dependencies for the Agentic AI Personal Agent System
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# LLM and AI dependencies
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.1
openai==1.3.7
anthropic==0.7.8

# Web scraping and data processing
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2
requests==2.31.0
pandas==2.1.4
numpy==1.25.2

# Email handling
smtplib2==0.2.1
email-validator==2.1.0
imaplib2==3.6

# Task management and scheduling
schedule==1.2.0
croniter==2.0.1
apscheduler==3.10.4

# Database and storage
sqlalchemy==2.0.23
sqlite3
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
rich==13.7.0
typer==0.9.0
pyyaml==6.0.1
jinja2==3.1.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0